fix(getversion): allow user to pass commitmessage parameter to generate new version without commit id
Bug fix for reliza-cli getversion command:
Before sourceCodeEntry info would only be passed along with gql request if the user supplied a commit id with the '--commmit' flag, but now sourceCodeEntry will also be passed with getversion gql request if '--commitmessage' flag is present, even if '--commit' flag is not.
This will allow new version to be generated according to parsed commit message without needed to supply commit ID, however it is important to note that a Source Code Entry will not be created for the release unless a commit ID is supplied with the '--commit' flag.